FoundrCard/requirements.txt
amqp==5.3.1
asgiref==3.8.1
beautifulsoup4==4.13.4
billiard==4.2.1
cachetools==5.5.2
celery==5.5.3
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
cryptography==45.0.4
decouple==0.0.7
dj-database-url==3.0.0
dj-rest-auth==7.0.1
Django==5.2.3
django-allauth==65.9.0
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-jazzmin==3.0.1
django-timezone-field==7.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
google==3.0.0
google-auth==2.40.3
gunicorn==23.0.0
h11==0.16.0
idna==3.10
kombu==5.5.4
packaging==25.0
pillow==11.2.1
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
redis==6.2.0
requests==2.32.4
rsa==4.9.1
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
stripe==12.3.0
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
uvicorn-worker==0.3.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0